+++ docs/CHANGELOG.md
@@ -1,3 +1,6 @@
+### Version 7.1.10
+####
+
 ### Version 7.1.9
 #### Fixed:
 - Potential error caused by moving rows/columns where: